+++ test/android_database_test.dart
@@ -2,7 +2,7 @@
 
 import 'package:flutter_test/flutter_test.dart';
 
-//import 'package:flutter_android/android_database.dart';
+import 'package:flutter_android/android_database.dart';
 
 void main() {
   //Cursor cursor;
@@ -28,6 +28,31 @@ void main() {
   });
 
   group("android_database.MatrixCursor", () {
-    // TODO
+    test("MatrixCursor.empty", () {
+      final Cursor cursor = MatrixCursor.empty();
+      expect(cursor, isEmpty);
+      expect(cursor.position, equals(-1));
+      expect(cursor.isBeforeFirst, isTrue);
+      for (var row in cursor) {} // ignore: unused_local_variable
+      expect(cursor.position, equals(0));
+      expect(cursor.isAfterLast, isTrue);
+    });
+    test("MatrixCursor.from", () {
+      final Cursor cursor = MatrixCursor.from(
+        columns: <String>['a', 'b', 'c'],
+        rows: <List<dynamic>>[[1, 2, 3], [4, 5, 6], [7, 8, 9]],
+      );
+      expect(cursor, isNotEmpty);
+      expect(cursor.position, equals(-1));
+      expect(cursor.isBeforeFirst, isTrue);
+      for (var row in cursor) {
+        expect(row, hasLength(3));
+        expect(row.keys, contains('a'));
+        expect(row.keys, contains('b'));
+        expect(row.keys, contains('c'));
+      }
+      expect(cursor.position, equals(3));
+      expect(cursor.isAfterLast, isTrue);
+    });
   });
 }