+++ CHANGELOG.md
@@ -7,7 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.2.1] - 2018-12-06
-### Added
+### Changed
+- Enhanced the example application and added a screenshot
 
 ## [0.2.0] - 2018-12-05
 ### Changed