+++ CHANGELOG.md
@@ -6,8 +6,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.2] - 2018-12-07
-### Added
+## [0.2.2] - 2018-12-08
+### Changed
+- Made `android_database.Cursor` iterable
 
 ## [0.2.1] - 2018-12-06
 ### Changed